Dev and test datasets, first real rule
Created a test and train dataset for writting and evaluating CFG rules.
The train dataset can be looked at, examined and based on that - the
rules can be written. The test dataset is only there for evaluation,
using it to write rules is forbidden.

Written first actual rule that checks if un|vai sides are pluraly equal.

Written unit tests that check if there is no false positive errors, also
written unit tests, which check the evaluation of the CFG rules.